Move errors one level up in the module hierarchy
