+++ lib/nemid.rb
@@ -2,6 +2,7 @@ require 'base64'
 require "nemid/version"
 require "nemid/authentication"
 require "nemid/crypto"
+require "nemid/errors"
 require "nemid/xmldsig"
 
 module NemID
