+++ lib/nemid/authentication.rb
@@ -3,6 +3,5 @@ module NemID
   end
 end
 
-require_relative "authentication/errors"
 require_relative "authentication/params"
 require_relative "authentication/response"
